Migrate sketeton and search screens to ts
