+++ components/screens/SearchChannelPage.tsx
@@ -1,28 +1,23 @@
-import Layout from 'components/common/Layout'
-import MainTitle from 'components/MainTitle'
 import { useRouter } from 'next/router'
 import { useFetch } from 'lib/hooks'
+import Layout from 'components/common/Layout'
+import MainTitle from 'components/MainTitle'
 import ClearSearchButton from 'components/search/ClearSearchButton'
-import EpisodeListContainer from 'components/episodes/EpisodeListContainer'
 import SearchErrorMessage from 'components/search/SearchErrorMessage'
-import { getEpisodesSearchUrl } from 'lib/constants'
+import { getChannelsSearchUrl } from 'lib/constants'
+import { ChannelsGrid } from 'components/channel/ChannelsGrid'
 
-export const SearchEpisodesPage = () => {
+export const SearchChannelPage = () => {
   const router = useRouter()
   const { keyword } = router.query
 
-  const episodesUrl = getEpisodesSearchUrl(keyword)
-
-  const { data, isLoading } = useFetch(episodesUrl, 'audio_clips')
+  const channelsUrl = getChannelsSearchUrl(keyword)
+  const { data, isLoading } = useFetch(channelsUrl, 'channels')
 
   const Content = () => (
     <>
       <MainTitle title={`"${keyword}" in search`} />
-      <EpisodeListContainer
-        title="All Episodes"
-        episodes={data}
-        loading={isLoading}
-      />
+      <ChannelsGrid title="All channels" channels={data} loading={isLoading} />
     </>
   )
 
@@ -34,7 +29,7 @@ export const SearchEpisodesPage = () => {
       button={<ClearSearchButton />}
     >
       {!isLoading && data.length === 0 ? (
-        <SearchErrorMessage keyword={keyword} />
+        <SearchErrorMessage keyword={keyword as string} />
       ) : (
         <Content />
       )}
