+++ components/screens/SearchEpisodesPage.tsx
@@ -1,23 +1,28 @@
-import { useRouter } from 'next/router'
-import { useFetch } from 'lib/hooks'
 import Layout from 'components/common/Layout'
 import MainTitle from 'components/MainTitle'
+import { useRouter } from 'next/router'
+import { useFetch } from 'lib/hooks'
 import ClearSearchButton from 'components/search/ClearSearchButton'
+import EpisodeListContainer from 'components/episodes/EpisodeListContainer'
 import SearchErrorMessage from 'components/search/SearchErrorMessage'
-import { getChannelsSearchUrl } from 'lib/constants'
-import { ChannelsGrid } from 'components/channel/ChannelsGrid'
+import { getEpisodesSearchUrl } from 'lib/constants'
 
-export const SearchChannelPage = () => {
+export const SearchEpisodesPage = () => {
   const router = useRouter()
   const { keyword } = router.query
 
-  const channelsUrl = getChannelsSearchUrl(keyword)
-  const { data, isLoading } = useFetch(channelsUrl, 'channels')
+  const episodesUrl = getEpisodesSearchUrl(keyword)
+
+  const { data, isLoading } = useFetch(episodesUrl, 'audio_clips')
 
   const Content = () => (
     <>
       <MainTitle title={`"${keyword}" in search`} />
-      <ChannelsGrid title="All channels" channels={data} loading={isLoading} />
+      <EpisodeListContainer
+        title="All Episodes"
+        episodes={data}
+        loading={isLoading}
+      />
     </>
   )
 
@@ -29,7 +34,7 @@ export const SearchChannelPage = () => {
       button={<ClearSearchButton />}
     >
       {!isLoading && data.length === 0 ? (
-        <SearchErrorMessage keyword={keyword} />
+        <SearchErrorMessage keyword={keyword as string} />
       ) : (
         <Content />
       )}
